+++ test/Dapper.AOT.Test.Integration/Setup/ExpressionParsingException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Dapper.AOT.Test.Integration.Setup;
+
+class ExpressionParsingException(IEnumerable<string> errors)
+    : Exception(string.Join(Environment.NewLine, errors))
+{
+}
